Revert "Memory functions implementation and wrapping"
